Better builds
This will slim down the Cake script, build with warnings-as-errors and
also remove some tooling and let Cake handle it instead. Try to improve
build times as well.

+++ src/picotorrent/ui/dialogs/preferenceslabelspage.cpp
@@ -241,7 +241,7 @@ void PreferencesLabelsPage::Save()
         m_cfg->UpsertLabel(*lbl);
     }
 
-    for (int i = 0; i < m_removedLabels.size(); i++)
+    for (size_t i = 0; i < m_removedLabels.size(); i++)
     {
         m_cfg->DeleteLabel(m_removedLabels.at(i));
     }
